Create Base Converter
This week's practice lab. We were instructed to convert a given integer from one base to another.  We were told to accept three integers, the first was a number representing the base we were to convert a given integer from, the second, representing the base we were converting to and the third, the given integer.